win/uwp split
