+++ Code/Engine/Foundation/Platform/Win/OSFile_Win.cpp
@@ -1,6 +1,6 @@
 #include <Foundation/FoundationPCH.h>
 
-#if EZ_ENABLED(EZ_PLATFORM_WINDOWS)
+#if EZ_ENABLED(EZ_PLATFORM_WINDOWS_DESKTOP)
 
 #  include <Foundation/IO/OSFile.h>
 #  include <Foundation/Logging/Log.h>
@@ -8,20 +8,7 @@
 #  include <Foundation/Strings/StringConversion.h>
 #  include <Foundation/Threading/ThreadUtils.h>
 
-// Defined in Timestamp_win.h
-ezInt64 FileTimeToEpoch(FILETIME fileTime);
-
-static ezUInt64 HighLowToUInt64(ezUInt32 uiHigh32, ezUInt32 uiLow32)
-{
-  ezUInt64 uiHigh64 = uiHigh32;
-  ezUInt64 uiLow64 = uiLow32;
-
-  return (uiHigh64 << 32) | uiLow64;
-}
-
-#  if EZ_ENABLED(EZ_PLATFORM_WINDOWS_DESKTOP)
-
-#    include <Shlobj.h>
+#  include <Shlobj.h>
 
 ezResult ezOSFile::InternalOpen(ezStringView sFile, ezFileOpenMode::Enum OpenMode, ezFileShareMode::Enum FileShareMode)
 {
@@ -197,7 +184,7 @@ ezUInt64 ezOSFile::InternalGetFilePosition() const
   long int uiHigh32 = 0;
   ezUInt32 uiLow32 = SetFilePointer(m_FileData.m_pFileHandle, 0, &uiHigh32, FILE_CURRENT);
 
-  return HighLowToUInt64(uiHigh32, uiLow32);
+  return ezMath::MakeUInt64(uiHigh32, uiLow32);
 }
 
 void ezOSFile::InternalSetFilePosition(ezInt64 iDistance, ezFileSeekMode::Enum Pos) const
@@ -289,273 +276,6 @@ ezResult ezOSFile::InternalMoveFileOrDirectory(ezStringView sDirectoryFrom, ezSt
   return EZ_SUCCESS;
 }
 
-#  endif
-
-ezResult ezOSFile::InternalGetFileStats(ezStringView sFileOrFolder, ezFileStats& out_Stats)
-{
-  ezStringBuilder s = sFileOrFolder;
-
-  // FindFirstFile does not like paths that end with a separator, so remove them all
-  s.Trim(nullptr, "/\\");
-
-  // handle the case that this query is done on the 'device part' of a path
-  if (s.GetCharacterCount() <= 2) // 'C:', 'D:', 'E' etc.
-  {
-    s.ToUpper();
-
-    out_Stats.m_uiFileSize = 0;
-    out_Stats.m_bIsDirectory = true;
-    out_Stats.m_sParentPath.Clear();
-    out_Stats.m_sName = s;
-    out_Stats.m_LastModificationTime = ezTimestamp::MakeInvalid();
-    return EZ_SUCCESS;
-  }
-
-  WIN32_FIND_DATAW data;
-  HANDLE hSearch = FindFirstFileW(ezDosDevicePath(s), &data);
-
-  if ((hSearch == nullptr) || (hSearch == INVALID_HANDLE_VALUE))
-    return EZ_FAILURE;
-
-  out_Stats.m_uiFileSize = HighLowToUInt64(data.nFileSizeHigh, data.nFileSizeLow);
-  out_Stats.m_bIsDirectory = (data.dwFileAttributes & FILE_ATTRIBUTE_DIRECTORY) != 0;
-  out_Stats.m_sParentPath = sFileOrFolder;
-  out_Stats.m_sParentPath.PathParentDirectory();
-  out_Stats.m_sName = data.cFileName;
-  out_Stats.m_LastModificationTime = ezTimestamp::MakeFromInt(FileTimeToEpoch(data.ftLastWriteTime), ezSIUnitOfTime::Microsecond);
-
-  FindClose(hSearch);
-  return EZ_SUCCESS;
-}
-
-#  if EZ_ENABLED(EZ_SUPPORTS_FILE_ITERATORS)
-
-ezFileSystemIterator::ezFileSystemIterator() = default;
-
-ezFileSystemIterator::~ezFileSystemIterator()
-{
-  while (!m_Data.m_Handles.IsEmpty())
-  {
-    FindClose(m_Data.m_Handles.PeekBack());
-    m_Data.m_Handles.PopBack();
-  }
-}
-
-bool ezFileSystemIterator::IsValid() const
-{
-  return !m_Data.m_Handles.IsEmpty();
-}
-
-void ezFileSystemIterator::StartSearch(ezStringView sSearchStart, ezBitflags<ezFileSystemIteratorFlags> flags /*= ezFileSystemIteratorFlags::All*/)
-{
-  EZ_ASSERT_DEV(m_Data.m_Handles.IsEmpty(), "Cannot start another search.");
-
-  m_sSearchTerm = sSearchStart;
-
-  ezStringBuilder sSearch = sSearchStart;
-  sSearch.MakeCleanPath();
-
-  // same as just passing in the folder path, so remove this
-  if (sSearch.EndsWith("/*"))
-    sSearch.Shrink(0, 2);
-
-  // The Windows documentation disallows trailing (back)slashes.
-  sSearch.Trim(nullptr, "/");
-
-  // Since the use of wildcard-ed file names will disable recursion, we ensure both are not used simultaneously.
-  const bool bHasWildcard = sSearch.FindLastSubString("*") || sSearch.FindLastSubString("?");
-  EZ_ASSERT_DEV(flags.IsSet(ezFileSystemIteratorFlags::Recursive) == false || bHasWildcard == false, "Recursive file iteration does not support wildcards. Either don't use recursion, or filter the filenames manually.");
-
-  if (!bHasWildcard && ezOSFile::ExistsDirectory(sSearch))
-  {
-    // when calling FindFirstFileW with a path to a folder (e.g. "C:/test") it will report "test" as the very first item
-    // which is typically NOT what one wants, instead you want items INSIDE that folder to be reported
-    // this is especially annoying when 'Recursion' is disabled, as "C:/test" would result in "C:/test" being reported
-    // but no items inside it
-    // therefore, when the start search points to a directory, we append "/*" to force the search inside the folder
-    sSearch.Append("/*");
-  }
-
-  m_sCurPath = sSearch.GetFileDirectory();
-
-  EZ_ASSERT_DEV(sSearch.IsAbsolutePath(), "The path '{0}' is not absolute.", m_sCurPath);
-
-  m_Flags = flags;
-
-  WIN32_FIND_DATAW data;
-  HANDLE hSearch = FindFirstFileW(ezDosDevicePath(sSearch), &data);
-
-  if ((hSearch == nullptr) || (hSearch == INVALID_HANDLE_VALUE))
-    return;
-
-  m_CurFile.m_uiFileSize = HighLowToUInt64(data.nFileSizeHigh, data.nFileSizeLow);
-  m_CurFile.m_bIsDirectory = (data.dwFileAttributes & FILE_ATTRIBUTE_DIRECTORY) != 0;
-  m_CurFile.m_sParentPath = m_sCurPath;
-  m_CurFile.m_sParentPath.TrimRight("/\\"); // remove trailing slashes
-  m_CurFile.m_sName = data.cFileName;
-  m_CurFile.m_LastModificationTime = ezTimestamp::MakeFromInt(FileTimeToEpoch(data.ftLastWriteTime), ezSIUnitOfTime::Microsecond);
-
-  m_Data.m_Handles.PushBack(hSearch);
-
-  if ((m_CurFile.m_sName == "..") || (m_CurFile.m_sName == "."))
-  {
-    Next(); // will search for the next file or folder that is not ".." or "." ; might return false though
-    return;
-  }
-
-  if (m_CurFile.m_bIsDirectory)
-  {
-    if (!m_Flags.IsSet(ezFileSystemIteratorFlags::ReportFolders))
-    {
-      Next();
-      return;
-    }
-  }
-  else
-  {
-    if (!m_Flags.IsSet(ezFileSystemIteratorFlags::ReportFiles))
-    {
-      Next();
-      return;
-    }
-  }
-}
-
-ezInt32 ezFileSystemIterator::InternalNext()
-{
-  constexpr ezInt32 ReturnFailure = 0;
-  constexpr ezInt32 ReturnSuccess = 1;
-  constexpr ezInt32 ReturnCallInternalNext = 2;
-
-  if (m_Data.m_Handles.IsEmpty())
-    return ReturnFailure;
-
-  if (m_Flags.IsSet(ezFileSystemIteratorFlags::Recursive) && m_CurFile.m_bIsDirectory && (m_CurFile.m_sName != "..") && (m_CurFile.m_sName != "."))
-  {
-    m_sCurPath.AppendPath(m_CurFile.m_sName);
-
-    ezStringBuilder sNewSearch = m_sCurPath;
-    sNewSearch.AppendPath("*");
-
-    WIN32_FIND_DATAW data;
-    HANDLE hSearch = FindFirstFileW(ezDosDevicePath(sNewSearch), &data);
-
-    if ((hSearch != nullptr) && (hSearch != INVALID_HANDLE_VALUE))
-    {
-      m_CurFile.m_uiFileSize = HighLowToUInt64(data.nFileSizeHigh, data.nFileSizeLow);
-      m_CurFile.m_bIsDirectory = (data.dwFileAttributes & FILE_ATTRIBUTE_DIRECTORY) != 0;
-      m_CurFile.m_sParentPath = m_sCurPath;
-      EZ_ASSERT_DEBUG(!m_CurFile.m_sParentPath.EndsWith("/") && !m_CurFile.m_sParentPath.EndsWith("\\"), "Unexpected path separator.");
-      m_CurFile.m_sName = data.cFileName;
-      m_CurFile.m_LastModificationTime = ezTimestamp::MakeFromInt(FileTimeToEpoch(data.ftLastWriteTime), ezSIUnitOfTime::Microsecond);
-
-      m_Data.m_Handles.PushBack(hSearch);
-
-      if ((m_CurFile.m_sName == "..") || (m_CurFile.m_sName == "."))
-        return ReturnCallInternalNext; // will search for the next file or folder that is not ".." or "." ; might return false though
-
-      if (m_CurFile.m_bIsDirectory)
-      {
-        if (!m_Flags.IsSet(ezFileSystemIteratorFlags::ReportFolders))
-          return ReturnCallInternalNext;
-      }
-      else
-      {
-        if (!m_Flags.IsSet(ezFileSystemIteratorFlags::ReportFiles))
-          return ReturnCallInternalNext;
-      }
-
-      return ReturnSuccess;
-    }
-
-    // if the recursion did not work, just iterate in this folder further
-  }
-
-  WIN32_FIND_DATAW data;
-  if (!FindNextFileW(m_Data.m_Handles.PeekBack(), &data))
-  {
-    // nothing found in this directory anymore
-    FindClose(m_Data.m_Handles.PeekBack());
-    m_Data.m_Handles.PopBack();
-
-    if (m_Data.m_Handles.IsEmpty())
-      return ReturnFailure;
-
-    m_sCurPath.PathParentDirectory();
-    if (m_sCurPath.EndsWith("/"))
-    {
-      m_sCurPath.Shrink(0, 1); // Remove trailing /
-    }
-
-    return ReturnCallInternalNext;
-  }
-
-  m_CurFile.m_uiFileSize = HighLowToUInt64(data.nFileSizeHigh, data.nFileSizeLow);
-  m_CurFile.m_bIsDirectory = (data.dwFileAttributes & FILE_ATTRIBUTE_DIRECTORY) != 0;
-  m_CurFile.m_sParentPath = m_sCurPath;
-  m_CurFile.m_sParentPath.TrimRight("/\\"); // remove trailing slashes
-  m_CurFile.m_sName = data.cFileName;
-  m_CurFile.m_LastModificationTime = ezTimestamp::MakeFromInt(FileTimeToEpoch(data.ftLastWriteTime), ezSIUnitOfTime::Microsecond);
-
-  if ((m_CurFile.m_sName == "..") || (m_CurFile.m_sName == "."))
-    return ReturnCallInternalNext;
-
-  if (m_CurFile.m_bIsDirectory)
-  {
-    if (!m_Flags.IsSet(ezFileSystemIteratorFlags::ReportFolders))
-      return ReturnCallInternalNext;
-  }
-  else
-  {
-    if (!m_Flags.IsSet(ezFileSystemIteratorFlags::ReportFiles))
-      return ReturnCallInternalNext;
-  }
-
-  return ReturnSuccess;
-}
-
-#  endif
-
-ezStringView ezOSFile::GetApplicationPath()
-{
-  if (s_sApplicationPath.IsEmpty())
-  {
-    ezUInt32 uiRequiredLength = 512;
-    ezHybridArray<wchar_t, 1024> tmp;
-
-    while (true)
-    {
-      tmp.SetCountUninitialized(uiRequiredLength);
-
-      // reset last error code
-      SetLastError(ERROR_SUCCESS);
-
-      const ezUInt32 uiLength = GetModuleFileNameW(nullptr, tmp.GetData(), tmp.GetCount() - 1);
-      const DWORD error = GetLastError();
-
-      if (error == ERROR_SUCCESS)
-      {
-        tmp[uiLength] = L'\0';
-        break;
-      }
-
-      if (error == ERROR_INSUFFICIENT_BUFFER)
-      {
-        uiRequiredLength += 512;
-        continue;
-      }
-
-      EZ_REPORT_FAILURE("GetModuleFileNameW failed: {0}", ezArgErrorCode(error));
-    }
-
-    s_sApplicationPath = ezStringUtf8(tmp.GetData()).GetData();
-  }
-
-  return s_sApplicationPath;
-}
-
-#  if EZ_ENABLED(EZ_PLATFORM_WINDOWS_DESKTOP)
-
 ezString ezOSFile::GetUserDataFolder(ezStringView sSubFolder)
 {
   if (s_sUserDataPath.IsEmpty())
@@ -626,27 +346,196 @@ ezString ezOSFile::GetUserDocumentsFolder(ezStringView sSubFolder /*= {}*/)
   return s;
 }
 
-#  endif
 
-const ezString ezOSFile::GetCurrentWorkingDirectory()
+#endif
+
+#if EZ_ENABLED(EZ_PLATFORM_WINDOWS) && EZ_ENABLED(EZ_SUPPORTS_FILE_ITERATORS)
+
+// Defined in Timestamp_win.h
+ezInt64 FileTimeToEpoch(FILETIME fileTime);
+
+ezFileSystemIterator::ezFileSystemIterator() = default;
+
+ezFileSystemIterator::~ezFileSystemIterator()
 {
-  const ezUInt32 uiRequiredLength = GetCurrentDirectoryW(0, nullptr);
-
-  ezHybridArray<wchar_t, 1024> tmp;
-  tmp.SetCountUninitialized(uiRequiredLength + 16);
-
-  if (GetCurrentDirectoryW(tmp.GetCount() - 1, tmp.GetData()) == 0)
+  while (!m_Data.m_Handles.IsEmpty())
   {
-    EZ_REPORT_FAILURE("GetCurrentDirectoryW failed: {}", ezArgErrorCode(GetLastError()));
-    return ezString();
+    FindClose(m_Data.m_Handles.PeekBack());
+    m_Data.m_Handles.PopBack();
+  }
+}
+
+bool ezFileSystemIterator::IsValid() const
+{
+  return !m_Data.m_Handles.IsEmpty();
+}
+
+void ezFileSystemIterator::StartSearch(ezStringView sSearchStart, ezBitflags<ezFileSystemIteratorFlags> flags /*= ezFileSystemIteratorFlags::All*/)
+{
+  EZ_ASSERT_DEV(m_Data.m_Handles.IsEmpty(), "Cannot start another search.");
+
+  m_sSearchTerm = sSearchStart;
+
+  ezStringBuilder sSearch = sSearchStart;
+  sSearch.MakeCleanPath();
+
+  // same as just passing in the folder path, so remove this
+  if (sSearch.EndsWith("/*"))
+    sSearch.Shrink(0, 2);
+
+  // The Windows documentation disallows trailing (back)slashes.
+  sSearch.Trim(nullptr, "/");
+
+  // Since the use of wildcard-ed file names will disable recursion, we ensure both are not used simultaneously.
+  const bool bHasWildcard = sSearch.FindLastSubString("*") || sSearch.FindLastSubString("?");
+  EZ_ASSERT_DEV(flags.IsSet(ezFileSystemIteratorFlags::Recursive) == false || bHasWildcard == false, "Recursive file iteration does not support wildcards. Either don't use recursion, or filter the filenames manually.");
+
+  if (!bHasWildcard && ezOSFile::ExistsDirectory(sSearch))
+  {
+    // when calling FindFirstFileW with a path to a folder (e.g. "C:/test") it will report "test" as the very first item
+    // which is typically NOT what one wants, instead you want items INSIDE that folder to be reported
+    // this is especially annoying when 'Recursion' is disabled, as "C:/test" would result in "C:/test" being reported
+    // but no items inside it
+    // therefore, when the start search points to a directory, we append "/*" to force the search inside the folder
+    sSearch.Append("/*");
   }
 
-  tmp[uiRequiredLength] = L'\0';
+  m_sCurPath = sSearch.GetFileDirectory();
 
-  ezStringBuilder clean = ezStringUtf8(tmp.GetData()).GetData();
-  clean.MakeCleanPath();
+  EZ_ASSERT_DEV(sSearch.IsAbsolutePath(), "The path '{0}' is not absolute.", m_sCurPath);
 
-  return clean;
+  m_Flags = flags;
+
+  WIN32_FIND_DATAW data;
+  HANDLE hSearch = FindFirstFileW(ezDosDevicePath(sSearch), &data);
+
+  if ((hSearch == nullptr) || (hSearch == INVALID_HANDLE_VALUE))
+    return;
+
+  m_CurFile.m_uiFileSize = ezMath::MakeUInt64(data.nFileSizeHigh, data.nFileSizeLow);
+  m_CurFile.m_bIsDirectory = (data.dwFileAttributes & FILE_ATTRIBUTE_DIRECTORY) != 0;
+  m_CurFile.m_sParentPath = m_sCurPath;
+  m_CurFile.m_sParentPath.TrimRight("/\\"); // remove trailing slashes
+  m_CurFile.m_sName = data.cFileName;
+  m_CurFile.m_LastModificationTime = ezTimestamp::MakeFromInt(FileTimeToEpoch(data.ftLastWriteTime), ezSIUnitOfTime::Microsecond);
+
+  m_Data.m_Handles.PushBack(hSearch);
+
+  if ((m_CurFile.m_sName == "..") || (m_CurFile.m_sName == "."))
+  {
+    Next(); // will search for the next file or folder that is not ".." or "." ; might return false though
+    return;
+  }
+
+  if (m_CurFile.m_bIsDirectory)
+  {
+    if (!m_Flags.IsSet(ezFileSystemIteratorFlags::ReportFolders))
+    {
+      Next();
+      return;
+    }
+  }
+  else
+  {
+    if (!m_Flags.IsSet(ezFileSystemIteratorFlags::ReportFiles))
+    {
+      Next();
+      return;
+    }
+  }
+}
+
+ezInt32 ezFileSystemIterator::InternalNext()
+{
+  constexpr ezInt32 ReturnFailure = 0;
+  constexpr ezInt32 ReturnSuccess = 1;
+  constexpr ezInt32 ReturnCallInternalNext = 2;
+
+  if (m_Data.m_Handles.IsEmpty())
+    return ReturnFailure;
+
+  if (m_Flags.IsSet(ezFileSystemIteratorFlags::Recursive) && m_CurFile.m_bIsDirectory && (m_CurFile.m_sName != "..") && (m_CurFile.m_sName != "."))
+  {
+    m_sCurPath.AppendPath(m_CurFile.m_sName);
+
+    ezStringBuilder sNewSearch = m_sCurPath;
+    sNewSearch.AppendPath("*");
+
+    WIN32_FIND_DATAW data;
+    HANDLE hSearch = FindFirstFileW(ezDosDevicePath(sNewSearch), &data);
+
+    if ((hSearch != nullptr) && (hSearch != INVALID_HANDLE_VALUE))
+    {
+      m_CurFile.m_uiFileSize = ezMath::MakeUInt64(data.nFileSizeHigh, data.nFileSizeLow);
+      m_CurFile.m_bIsDirectory = (data.dwFileAttributes & FILE_ATTRIBUTE_DIRECTORY) != 0;
+      m_CurFile.m_sParentPath = m_sCurPath;
+      EZ_ASSERT_DEBUG(!m_CurFile.m_sParentPath.EndsWith("/") && !m_CurFile.m_sParentPath.EndsWith("\\"), "Unexpected path separator.");
+      m_CurFile.m_sName = data.cFileName;
+      m_CurFile.m_LastModificationTime = ezTimestamp::MakeFromInt(FileTimeToEpoch(data.ftLastWriteTime), ezSIUnitOfTime::Microsecond);
+
+      m_Data.m_Handles.PushBack(hSearch);
+
+      if ((m_CurFile.m_sName == "..") || (m_CurFile.m_sName == "."))
+        return ReturnCallInternalNext; // will search for the next file or folder that is not ".." or "." ; might return false though
+
+      if (m_CurFile.m_bIsDirectory)
+      {
+        if (!m_Flags.IsSet(ezFileSystemIteratorFlags::ReportFolders))
+          return ReturnCallInternalNext;
+      }
+      else
+      {
+        if (!m_Flags.IsSet(ezFileSystemIteratorFlags::ReportFiles))
+          return ReturnCallInternalNext;
+      }
+
+      return ReturnSuccess;
+    }
+
+    // if the recursion did not work, just iterate in this folder further
+  }
+
+  WIN32_FIND_DATAW data;
+  if (!FindNextFileW(m_Data.m_Handles.PeekBack(), &data))
+  {
+    // nothing found in this directory anymore
+    FindClose(m_Data.m_Handles.PeekBack());
+    m_Data.m_Handles.PopBack();
+
+    if (m_Data.m_Handles.IsEmpty())
+      return ReturnFailure;
+
+    m_sCurPath.PathParentDirectory();
+    if (m_sCurPath.EndsWith("/"))
+    {
+      m_sCurPath.Shrink(0, 1); // Remove trailing /
+    }
+
+    return ReturnCallInternalNext;
+  }
+
+  m_CurFile.m_uiFileSize = ezMath::MakeUInt64(data.nFileSizeHigh, data.nFileSizeLow);
+  m_CurFile.m_bIsDirectory = (data.dwFileAttributes & FILE_ATTRIBUTE_DIRECTORY) != 0;
+  m_CurFile.m_sParentPath = m_sCurPath;
+  m_CurFile.m_sParentPath.TrimRight("/\\"); // remove trailing slashes
+  m_CurFile.m_sName = data.cFileName;
+  m_CurFile.m_LastModificationTime = ezTimestamp::MakeFromInt(FileTimeToEpoch(data.ftLastWriteTime), ezSIUnitOfTime::Microsecond);
+
+  if ((m_CurFile.m_sName == "..") || (m_CurFile.m_sName == "."))
+    return ReturnCallInternalNext;
+
+  if (m_CurFile.m_bIsDirectory)
+  {
+    if (!m_Flags.IsSet(ezFileSystemIteratorFlags::ReportFolders))
+      return ReturnCallInternalNext;
+  }
+  else
+  {
+    if (!m_Flags.IsSet(ezFileSystemIteratorFlags::ReportFiles))
+      return ReturnCallInternalNext;
+  }
+
+  return ReturnSuccess;
 }
 
 #endif
